api/requirements.txt
alembic==1.9.2
aniso8601==9.0.1
attrs==22.2.0
certifi==2022.12.7
chardet==4.0.0
click==8.1.3
dnspython==1.16.0
EasyProcess==0.3
Flask==1.1.2
flask-accepts==0.17.6
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
flask-restplus==0.13.0
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SocketIO==4.3.2
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
loguru==0.5.3
lxml==4.6.2
Mako==1.1.4
markdown2==2.4.0
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.18.0
ply==3.11
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
pycryptodomex==3.10.1
pyjon.utils==0.7
pyrsistent==0.17.3
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2022.7.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
tzlocal==2.1
urllib3==1.26.3
Werkzeug==0.16.0
zipp==3.11.0
